### 2020-03-06 #2146 @whedon
[PRE REVIEW]: Sensie: Probing the sensitivity of neural networks
**Submitting author:** @coljac (<a href="http://orcid.org/0000-0003-4239-4055">Colin Jacobs</a>)
**Repository:** <a href="https://github.com/coljac/sensie" target ="_blank">https://github.com/coljac/sensie</a>
**Version:** 0.1.0
