### 2020-04-30 #2146 @ejhigson
@gkthiruvathukal @arfon I would be happy to review this submission if you would like me to? (I have a lot of experience with neural networks and Python)
