### 2020-04-30 #2180 @whedon
[REVIEW]: Sensie: Probing the sensitivity of neural networks
**Submitting author:** @coljac (<a href="http://orcid.org/0000-0003-4239-4055">Colin Jacobs</a>)
**Repository:** <a href="https://github.com/coljac/sensie" target ="_blank">https://github.com/coljac/sensie</a>
**Version:** v1.0.0
