### 2020-04-30 #2180 @arfon
@ejhigson - thanks again for agreeing to review this submission for JOSS.

Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Any questions/concerns please let myself or @gkthiruvathukal know.
