### 2020-04-30 #2180 @arfon
:wave: @nilesh-patil @sssomani @chekoduadarsh @omshinde - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The package being reviewed here is [Sensie](https://github.com/coljac/sensie), a Python package for probing the sensitivity of neural networks.
