### 2020-04-30 #2180 @omshinde
Hi @arfon ! Thanks for consideration. Yes, I would be interested to review this submission for JOSS. This would be my maiden review for JOSS and I foresee a great learning expereince ahead. Please let  me know of the further steps. Thanks!
