### 2020-05-03 #2180 @ejhigson
Hi @coljac - congratulations on a really neat piece of software. I definitely see the usefulness of this package!

I took an initial look over it and have a few suggestions:

1. For the installation, I suggest adding your requirements to ``setup.py`` using the ``install_requires`` property so they are installed automatically (in addition to having them in the requirements.txt). You also may want to list your package on PyPI so users can install it with pip without doing a git clone.

2. Forgive me if I have missed them but I can't see any instructions for how to run your automated tests in the docs? When I try with ``nosetests`` I get some errors but maybe I am doing something wrong?

3. I would suggest adding a bit more detail to your community guidelines at https://sensie.readthedocs.io/en/latest/index.html. The review criteria is "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". Some of this is covered in your "Contact" section but maybe you can add more.

4. In the paper, you survey the literature on testing DNN sensitivity but you don't mention if there are any other software packages for doing this similar to ``sense``. Please could you add this? If the answer is no then its worth saying so explicitly.

5. One of the references in the paper, "Layer-Wise Relevance Propagation for Neural Networks with Local Renormalization Layers", is missing a DOI.
