### 2020-05-05 6309a1d
Changes for JOSS review; test but with credible interval method name change.
