+++ paper/paper.bib
@@ -59,6 +59,7 @@
   year = {2016},
   pages = {63--71},
   publisher = {{Springer}},
+  doi = {10.1007/978-3-319-44781-0_8}
   file = {/home/coljac/Zotero/storage/7TLNISSD/Binder et al. - 2016 - Layer-wise relevance propagation for neural networ.pdf;/home/coljac/Zotero/storage/TRFVPTUW/layerwise.pdf;/home/coljac/Zotero/storage/IWEL5W8K/978-3-319-44781-0_8.html}
 }
 
