### 2020-05-05 #2180 @coljac
Hi @ejhigson,

Thanks for taking the time to review Sensie, it's much appreciated. I have responded to your initial feedback as follows:

>    For the installation, I suggest adding your requirements to setup.py using the install_requires property so they are installed automatically (in addition to having them in the requirements.txt). You also may want to list your package on PyPI so users can install it with pip without doing a git clone.

Thanks for this suggestion. I have updated setup.py accordingly, and adjusted the documentation to make the installation procedure clearer.

>    Forgive me if I have missed them but I can't see any instructions for how to run your automated tests in the docs? When I try with nosetests I get some errors but maybe I am doing something wrong?

`pip install pytest` and then `pytest test` from the repo root should run the tests successfully.

I have also indicated this in the documentation:

"Optionally, install `pytest` with `pip install pytest`, then run the tests with `pytest test` from the repository root."

>    I would suggest adding a bit more detail to your community guidelines at https://sensie.readthedocs.io/en/latest/index.html. The review criteria is "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". Some of this is covered in your "Contact" section but maybe you can add more.

I have updated the documentation to read as follows on readthedocs and GitHub:

"Issues, Questions and Contributions
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

Any problems or questions? Email colin@coljac.net, or open an issue on GitHub at https://github.com/coljac/sensie.

Contributions are welcome and encouraged. Fork the GitHub repository to your own machine, make some changes, and push your work back up to the fork and open a pull request so that I can review the changes."

>    In the paper, you survey the literature on testing DNN sensitivity but you don't mention if there are any other software packages for doing this similar to sense. Please could you add this? If the answer is no then its worth saying so explicitly.

I'm not aware of anything that does what Sensie does precisely. I added the following to the paper:

"Several quality software packages exist to probe ANN internals, such as ``Innvestigate`, ``tf-explain``, and ``keras-vis``. These packages are geared towards traditional image-based applications, and cannot explore relationships between the neural network performance and arbitrary properties of the examples presented to the network. ``Sensie`` is designed to fill that gap."

>    One of the references in the paper, "Layer-Wise Relevance Propagation for Neural Networks with Local Renormalization Layers", is missing a DOI.

Fixed.
