### 2020-05-05 9fea952
Reverted plot function name; adjusted example to current API
