### 2020-05-07 #2180 @ejhigson
Hi @coljac - thank you very much for these changes!

The install and tests work perfectly for me, and the community guidelines additions look good

> "Several quality software packages exist to probe ANN internals, such as `Innvestigate`, `tf-explain`, and `keras-vis`. These packages are geared towards traditional image-based applications, and cannot explore relationships between the neural network performance and arbitrary properties of the examples presented to the network. `Sensie` is designed to fill that gap."

Thank you for adding this to the paper - I think it is very useful background information.

Rereading the paper I spotted a couple of minor typos: you are missing a comma after ``Innvestigate`` and in the caption of Figure 1 I think "ann" should be "an". Aside from this I think the paper looks great and is very well written and clear. I particularly like the examples in the paper.

Regarding documentation: this looks mostly complete but there are a few methods missing documentation as can be seen at https://sensie.readthedocs.io/en/latest/sensie.html and in the source code. Please could you check it over and fill in the missing methods' documentation?
