### 2020-05-07 #2180 @ejhigson
Also, looking at the examples on the jupyter notebooks, I think "MNIST example" is missing an
``import numpy as np``
statement in the first cell? With this addition it runs perfectly for me. "CIFAR10 example" also works perfectly for me. However "VGG example" seems to need some files from your local drive at ``/home/coljac/Downloads/imagenet/gorillas``(?) Is this example notebook meant to be runnable by other people?
