### 2020-05-07 #2180 @gkthiruvathukal
@arfon Just letting you know, with this comment, that I am resuming editorial activities and will see this and other submissions through the process. It's good to see @ejhigson has been offering good feedback for @coljac, who is working to address the issues.
