### 2020-05-08 89a23a4
Small changes re JOSS review
