### 2020-05-08 #2180 @coljac
Hi @ejhigson,

Thanks again for the follow up comments - and the bug/typo finding.

- I added the missing numpy import in the notebook example.
- I fixed the typos indicated.

Regarding the VGG example, for copyright reasons I didn't check in/make available the collection of gorilla pictures I used for this example, and so I've left that as an exercise for the reader (though they can of course see how things work from the state of the notebook checked in). I've added a note to this effect to the top. Let me know if you think this is an acceptable solution; otherwise, I can hunt down some images (in a different class) whose copyright I can vouch for.

As for the documentation, I ensured there are docstrings for all the methods and they show up on readthedocs; I also put in docstrings for all the underscore-methods for those accessing the source. That should all be in order now.

Thanks again for the comments and the attention to detail.
