### 2020-05-08 #2180 @ejhigson
@coljac thank you very much for your latest changes - they all look good to me! I think the VGG example now works well with the note at the top you just added telling people they can use their own images instead.

I am satisfied that you have met all the JOSS criteria - the paper is well written and the performance all seems as expected. Congratulations on a very neat software package and a nice paper!
