### 2020-05-08 #2180 @arfon
> @arfon Is it easy to add a reviewer after review has started? I seem to recall that wasn't easy to do in the past. I'll try to find a second. Let's hope more people are becoming available as some of the COVID-19 measures begin to lift.

You can do `@whedon add xxx as reviewer` here which will update the assignments but unfortunately you'll need to manually create a checklist for the second reviewer by editing the issue body at the top of this thread.
