### 2020-05-30 #2180 @omshinde
Hi @coljac!
cc: @gkthiruvathukal , @arfon

System settings while reviewing:
1. Installation - Local Machine, Google Colab
2. Running example notebooks - Google Colab

Thank you for the great software, I foresee Sensie as a great framework for sensitivity analysis of a machine learning/deep learning model. Please find my detailed thoughts below regarding the points mentioned in the review checklist:

1. Installation: Does installation proceed as outlined in the documentation?
 - The installation works perfectly for `python` version `>=3.5.4` but not for a version `<3.5.4`. I feel a note about this could be added in the installation instructions. Implicitly, this is happening because of the dependency arising from the usage of `pymc3` and hence, installing `Sensie` for a `python` version `<3.5.4` generates following output - ```ERROR: Package 'pymc3' requires a different Python: 3.5.2 not in '>=3.5.4'```. This would be helpful while creating the virtual environment specifying a python version `>=3.5.4`.

2. Documentation:
2.1 Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- The examples worked like charm. I had to add `import matplotlib.pyplot as plt` before plotting the histogram in the CIFAR10 example notebook.

2.2 Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
- The [Sensie Documentation](https://sensie.readthedocs.io/en/latest/) does not mentions anything about the automated tests. Though, information about `tests` is present in the [installation](https://github.com/coljac/sensie#installation) section of the project repository README but not in the documentation. Please verify if I missed anything.

3. Software paper
3.1 The paper is very well drafted and includes relevant information for using the software package. I could only see a difference in the title mentioned in the paper and the title mentioned in the documentation.
Paper Title - Sensie: Probing the sensitivity of neural networks
Documentation Title - Sensie: Simple neural network sensitivity

Although, I don't feel any reasons not to use both the titles but I am bringing this to the author's interest if they missed it by any chance.

3.2 The [Discussion and conclusion](https://github.com/coljac/sensie/blob/master/paper/paper.md#discussion-and-conclusion) section in the paper refers to Figure 1 but the images are not annotated with figure title.
Citing the content from the paper - "Figure 1 shows two examples; on the left, an output plot from Sensie showing the sensitivity of a trained model to rotation of the input images".
I think it would be a good idea to add figure title below the images to avoid any confusion.

4. Other
-  I found all the examples for models trained using `keras` framework. The documentation also mentions that with user supplied predictor function any framework should be usable. If the authors find it useful, they can explore using [ONNX](https://github.com/onnx/onnx) format for the model object and use the [onnx-converters](https://github.com/onnx/tutorials#converting-to-onnx-format) for using model objects from multiple frameworks. Well, these are my thoughts which I see as potential add-ons for future versions of `Sensie` on authors discretion.

- There are some spelling errors and typos in the repository. Please find them below:
1. [Calculation of the **signficance** of the effect](https://github.com/coljac/sensie#calculation-of-the-signficance-of-the-effect) -> [Calculation of the **significance** of the effect](https://github.com/coljac/sensie#calculation-of-the-signficance-of-the-effect)

2. [Installation](https://github.com/coljac/sensie#installation)
Which will **isntall** Sensie -> Which will **install** Sensie

My remarks:
Reproducibility in machine learning research is an important pursuit in the current research setting. Be it a scenario for deploying an ML/DL model on the production or for using the model for research on a different dataset, sensitivity analysis of the pre-trained model plays a significant role in understanding the variation of evaluation metrics with the data and model parameters. `Sensie` provides a great utility in performing the sensitivity of a machine learning model and quantifying changes in the performance of the model with respect to a parameter. I wish to congratulate the authors for their efforts and for providing a neat software package to the scientific community. Thanks!
Kind regards,
Rajat
