### 2020-06-01 e8538c4
Review comments - minor changes
