### 2020-06-01 #2180 @coljac
Hi @omshinde,

Thanks for the remarks and careful attention to the project.

I have addressed your comments as follows:

1. Python version

I altered the README and docs to indicate python 3.6 and above is required (there are f-strings in the Sensie code).

2. Documentation

- I added the matplotlib import to the example notebook.
- I added the reference to pytest and how to run the tests to the documentation.

3. Paper

Thanks for comments about the title - I settled on " Sensie: Probing the sensitivity of neural networks".

Regarding Figure 1 in the paper, you wrote "the images are not annotated with figure title". However, there is a *caption* ("Figure 1: Left: ....") immediately below the image. Do you mean that it would be better to alter the images themselves to add the name of the figure? That's a little non-standard, so I'm not clear on that.

I re-generated the PDF (see immediately above this comment); is it possible you were looking at the markdown in the repository? The github previewer doesn't render the captions properly, but they show up in the PDF.

I fixed the typos. Thanks for spotting those.

Thanks again for reviewing the code.

-Colin
