### 2020-06-01 #2180 @omshinde
Hi @coljac!

Thank you for addressing the comments.

> Regarding Figure 1 in the paper, you wrote "the images are not annotated with figure title". However, there is a caption ("Figure 1: Left: ....") immediately below the image. Do you mean that it would be better to alter the images themselves to add the name of the figure? That's a little non-standard, so I'm not clear on that.
>
> I re-generated the PDF (see immediately above this comment); is it possible you were looking at the markdown in the repository? The github previewer doesn't render the captions properly, but they show up in the PDF.

Oh, I missed out on this. Thanks for pointing it out and sorry for the inconvenience. Yes, the generated PDF looks perfect with the image properly captioned in the paper. Please keep it as it is.

I am completely satisfied with the `Sensie` software package and paper with regards to the JOSS review-checklist. Thanks again for your contribution and congratulations!

-Rajat
