### 2020-06-17 #2180 @Kevin-Mattheus-Moerman
@gkthiruvathukal I've assigned the archive DOI and also update the version label (pumped to v1.0.0). I also check the DOI metadata which seems okay. If you are happy to tick your boxes :point_up: and proceed, let me know. Thanks.
