### 2020-06-18 #2180 @Kevin-Mattheus-Moerman
@coljac some editorial comments on the paper:

- [ ] Check/fix double occurrence of `directly` in: `Some well-known methods for visualising and interpreting the outputs of DNNs include directly inspecting the learned features of a model directly`
- [ ] Please define the acronym ANN (or should it be DNN?)
- [ ] If the software you cite in footnotes has papers to cite please do add them.
- [ ] The paper is about to be processed for acceptance in JOSS. Please proofread the paper yourself again. In particular check the author names, affiliations, and acknowledgement section.

Once you've made these changes please call `@whedon generate pdf` to update the paper (use `@whedon check references` if you add new citations).
