### 2020-06-18 6fae370
Added references for other software.
