+++ paper/paper.bib
@@ -196,4 +196,29 @@
   keywords = {convnets,machine learning,toread}
 }
 
+@misc{raghakotkerasvis,
+  title={keras-vis},
+  author={Kotikalapudi, Raghavendra and contributors},
+  year={2017},
+  publisher={GitHub},
+  howpublished={\url{https://github.com/raghakot/keras-vis}},
+}
+
+@article{alberINNvestigateNeuralNetworks2018,
+  title = {{{iNNvestigate}} Neural Networks!},
+  author = {Alber, Maximilian and Lapuschkin, Sebastian and Seegerer, Philipp and H{\"a}gele, Miriam and Sch{\"u}tt, Kristof T. and Montavon, Gr{\'e}goire and Samek, Wojciech and M{\"u}ller, Klaus-Robert and D{\"a}hne, Sven and Kindermans, Pieter-Jan},
+  year = {2018},
+  month = aug,
+  url = {http://arxiv.org/abs/1808.04260},
+  urldate = {2019-10-23},
+  abstract = {In recent years, deep neural networks have revolutionized many application domains of machine learning and are key components of many critical decision or predictive processes. Therefore, it is crucial that domain specialists can understand and analyze actions and pre- dictions, even of the most complex neural network architectures. Despite these arguments neural networks are often treated as black boxes. In the attempt to alleviate this short- coming many analysis methods were proposed, yet the lack of reference implementations often makes a systematic comparison between the methods a major effort. The presented library iNNvestigate addresses this by providing a common interface and out-of-the- box implementation for many analysis methods, including the reference implementation for PatternNet and PatternAttribution as well as for LRP-methods. To demonstrate the versatility of iNNvestigate, we provide an analysis of image classifications for variety of state-of-the-art neural network architectures.},
+  archivePrefix = {arXiv},
+  eprint = {1808.04260},
+  eprinttype = {arxiv},
+  file = {/home/coljac/Zotero/storage/IT36VRJ6/Alber et al. - 2018 - iNNvestigate neural networks!.pdf;/home/coljac/Zotero/storage/M2EJG9GQ/1808.html},
+  journal = {arXiv:1808.04260 [cs, stat]},
+  keywords = {Computer Science - Machine Learning,Statistics - Machine Learning},
+  primaryClass = {cs, stat}
+}
+
 
