+++ paper/paper.md
@@ -19,11 +19,11 @@ bibliography: paper.bib
 
 # Introduction 
 
-Deep neural networks (DNNs) are finding increasing application across a wide variety of fields, including in industry and scientific research. Although DNNs are able to successfully tackle data problems that proved intractable to other methods, for instance in computer vision, they suffer from a lack of interpretability.  Some well-known methods for visualising and interpreting the outputs of DNNs include directly inspecting the learned features of a model directly (e.g. convolutional kernels and activation maps); or producing "saliency maps" which allow a human researcher to inspect which parts of an input (such as an image) played the most crucial role in the model reaching a particular determination. These algorithms include occluding parts of an image [@zeilerVisualizingUnderstandingConvolutional2014], guided backpropagation [@simonyanDeepConvolutionalNetworks2013], and more complex algorithms such as SmoothGrad [@smilkovSmoothGradRemovingNoise2017] and Layer-wise relevance propagation [@binderLayerwiseRelevancePropagation2016].
+Deep neural networks (DNNs) are finding increasing application across a wide variety of fields, including in industry and scientific research. Although DNNs are able to successfully tackle data problems that proved intractable to other methods, for instance in computer vision, they suffer from a lack of interpretability.  Some well-known methods for visualising and interpreting the outputs of DNNs include directly inspecting the learned features of a model (e.g. convolutional kernels and activation maps); or producing "saliency maps" which allow a human researcher to inspect which parts of an input (such as an image) played the most crucial role in the model reaching a particular determination. These algorithms include occluding parts of an image [@zeilerVisualizingUnderstandingConvolutional2014], guided backpropagation [@simonyanDeepConvolutionalNetworks2013], and more complex algorithms such as SmoothGrad [@smilkovSmoothGradRemovingNoise2017] and Layer-wise relevance propagation [@binderLayerwiseRelevancePropagation2016].
 
 However, simply inspecting the most salient regions of an input image (or other input) may not always be sufficiently interpretable. Here we present ``Sensie``, a python package to quickly inspect and quantify the sensitivity of a trained model to user-specified properties of the input domain, or under the arbitrary transformation of a test set.
 
-Several quality software packages exist to probe ANN internals, such as ``Innvestigate``[^innvestigate], ``tf-explain``[^tfe], and ``keras-vis``[^kvis]. These packages are geared towards traditional image-based applications, and cannot explore relationships between the neural network performance and arbitrary properties of the examples presented to the network. ``Sensie`` is designed to fill that gap.
+Several quality software packages exist to probe artificial neural network internals, such as ``Innvestigate`` [@alberINNvestigateNeuralNetworks2018], ``tf-explain``[^tfe], and ``keras-vis``[@raghakotkerasvis]. These packages are geared towards traditional image-based applications, and cannot explore relationships between the neural network performance and arbitrary properties of the examples presented to the network. ``Sensie`` is designed to fill that gap.
 
 # Method
 
@@ -66,7 +66,5 @@ The author acknowledges support from Karl Glazebrook’s Australian Research Cou
 # References
 
 [^bayesian]: See https://docs.pymc.io/notebooks/GLM-linear.html for an example using the PyMC3 probabilistic programming framework.
-[^innvestigate]: https://github.com/albermax/innvestigate
 [^tfe]: https://github.com/sicara/tf-explain
-[^kvis]: https://github.com/raghakot/keras-vis
 
