### 2020-06-18 #2180 @coljac
@Kevin-Mattheus-Moerman

Done and done. Thanks for catching a couple of issues.
