### 2020-06-19 #2180 @Kevin-Mattheus-Moerman
@coljac great I ticked all those boxes. Can you check that last potentially missing DOI? Is it this paper: https://ieeexplore.ieee.org/document/8237336 ?
