+++ paper/paper.bib
@@ -138,19 +138,6 @@
   number = {1}
 }
 
-@inproceedings{selvarajuGradCAMVisualExplanations2017,
-  title = {Grad-{{CAM}}: {{Visual Explanations From Deep Networks}} via {{Gradient}}-{{Based Localization}}},
-  shorttitle = {Grad-{{CAM}}},
-  booktitle = {Proceedings of the {{IEEE International Conference}} on {{Computer Vision}}},
-  author = {Selvaraju, Ramprasaath R. and Cogswell, Michael and Das, Abhishek and Vedantam, Ramakrishna and Parikh, Devi and Batra, Dhruv},
-  year = {2017},
-  pages = {618--626},
-  url = {http://openaccess.thecvf.com/content_iccv_2017/html/Selvaraju_Grad-CAM_Visual_Explanations_ICCV_2017_paper.html},
-  doi = {10.1109/ICCV.2017.74},
-  urldate = {2019-07-16},
-  file = {/home/coljac/Zotero/storage/F7W43JBJ/Selvaraju et al. - 2017 - Grad-CAM Visual Explanations From Deep Networks v.pdf;/home/coljac/Zotero/storage/EBU4PV83/Selvaraju_Grad-CAM_Visual_Explanations_ICCV_2017_paper.html}
-}
-
 @article{simonyanDeepConvolutionalNetworks2013,
   title = {Deep inside Convolutional Networks: {{Visualising}} Image Classification Models and Saliency Maps},
   shorttitle = {Deep inside Convolutional Networks},
