### 2020-06-19 #2180 @coljac
@Kevin-Mattheus-Moerman I fixed the erroneous DOI before realising I removed the reference from the body of the paper. I've updated accordingly.
